+++ robotidy/cli.py
@@ -290,6 +290,13 @@ def print_transformers_list():
     show_default=True,
 )
 @click.option(
+    "--line-length",
+    default=120,
+    type=int,
+    help="Max allowed characters per line",
+    show_default=True,
+)
+@click.option(
     "--list",
     "-l",
     is_eager=True,
@@ -338,6 +345,7 @@ def cli(
     separator: Optional[str],
     startline: Optional[int],
     endline: Optional[int],
+    line_length: int,
     list: bool,
     desc: Optional[str],
     output: Optional[Path],
@@ -376,6 +384,7 @@ def cli(
         start_line=startline,
         separator=separator,
         end_line=endline,
+        line_length=line_length,
     )
     tidy = Robotidy(
         transformers=transform,