+++ robotidy/transformers/__init__.py
@@ -38,6 +38,7 @@ TRANSFORMERS = [
     "RenameTestCases",
     "RenameKeywords",
     "ReplaceReturns",
+    "InlineIf",
 ]