+++ tests/utest/test_utils.py
@@ -10,11 +10,7 @@ from robotidy.utils import decorate_diff_with_color, split_args_from_name_or_pat
 @pytest.fixture
 def app():
     formatting_config = GlobalFormattingConfig(
-        space_count=4,
-        line_sep="auto",
-        start_line=None,
-        separator="space",
-        end_line=None,
+        space_count=4, line_sep="auto", start_line=None, separator="space", end_line=None, line_length=120
     )
     return Robotidy(
         transformers=[],
@@ -99,10 +95,6 @@ class TestUtils:
     def test_get_line_ending(self, line_sep, source_file, expected, app):
         source = str(Path(__file__).parent / "testdata" / "auto_line_sep" / source_file)
         app.formatting_config = GlobalFormattingConfig(
-            space_count=4,
-            line_sep=line_sep,
-            start_line=None,
-            separator="space",
-            end_line=None,
+            space_count=4, line_sep=line_sep, start_line=None, separator="space", end_line=None, line_length=120
         )
         assert app.get_line_ending(source) == expected